task3

+++ 0x0D-preprocessor/3-function_like_macro.h
@@ -1,5 +1,5 @@
 #ifndef FUNCTION_H
 #define FUNCTION_H
-#define ABS(x) (((((x) < 0) * -2) + 1) * x)
+#define ABS(x) (int)(((((x) < 0) * -2) + 1) * x)
 
 #endif
